+++ public/components/v2/create_job/job_stages/index_selection.js
@@ -14,7 +14,7 @@ import PropTypes from 'prop-types';
 
 import moment from 'moment';
 
-import { GenericRequest } from '../../../utils/requests';
+import { GenericRequest } from '@/services';
 
 export class IndexSelection extends Component {
   constructor(props) {
